stylematte/requirements.txt
albumentations==1.3.0
legacy==0.1.6
numpy==1.23.5
omegaconf==2.2.3
Pillow==9.4.0
scikit_image==0.19.3
scipy==1.10.0
torch
torchvision
tqdm==4.64.1
transformers==4.26.1
